fix create order

+++ project/app/Models/User.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use App\DTO\CartDTO;
 use App\DTO\OrderInfoDTO;
+use App\Exceptions\Cart\CartEmptyException;
 use App\Exceptions\Cart\CartProductLimitException;
 use App\Exceptions\Cart\CartProductOutStockException;
 use App\Exceptions\Cart\DuplicateCartProductException;
@@ -15,6 +16,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Laravel\Sanctum\HasApiTokens;
 
@@ -210,21 +212,27 @@ class User extends Authenticatable
     /**
      * @param OrderInfoDTO $order_info
      * @return void
+     * @throws CartEmptyException
      */
     public function createOrder(OrderInfoDTO $order_info): void
     {
-        /** @var Order $order */
-        $order = $this->orders()->create([
-            'address_id' => $order_info->address_id,
-            'phone' => $order_info->phone,
-            'cost' => $this->getCostCart(),
-            'email' => $order_info->email,
-            'delivery_time' => $order_info->delivery_time,
-        ]);
-        $products = $this->cart()->where('in_stock', true)->get();
-        foreach ($products as $product) {
-            $order->products()->attach($product->id, ['price' => $product->price, 'count' => $product->pivot->count]);
-            $this->cart()->detach($product->id);
+        if (!$this->cart()->where('in_stock', true)->count()) {
+            throw new CartEmptyException();
         }
+        DB::transaction(function () use ($order_info) {
+            /** @var Order $order */
+            $order = $this->orders()->create([
+                'address_id' => $order_info->address_id,
+                'phone' => $order_info->phone,
+                'cost' => $this->getCostCart(),
+                'email' => $order_info->email,
+                'delivery_time' => $order_info->delivery_time,
+            ]);
+            $products = $this->cart()->where('in_stock', true)->get();
+            foreach ($products as $product) {
+                $order->products()->attach($product->id, ['price' => $product->price, 'count' => $product->pivot->count]);
+                $this->cart()->detach($product->id);
+            }
+        }, 3);
     }
 }
